requirements.txt
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
chromedriver-binary==87.0.4280.20.0
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.17.0
lxml==4.6.2
MarkupSafe==1.1.1
python-dotenv==0.15.0
requests==2.25.0
schedule==0.6.0
selenium==3.141.0
soupsieve==2.0.1
SQLAlchemy==1.3.22
urllib3==1.26.2
Werkzeug==1.0.1
